+++ dflib-echarts/src/main/java/org/dflib/echarts/ECharts.java
@@ -1,5 +1,10 @@
 package org.dflib.echarts;
 
+import org.dflib.echarts.render.util.ElementIdGenerator;
+
+import java.io.File;
+import java.nio.file.Path;
+
 /**
  * An entry point to EChart builder. The builder would create HTML/JS code that renders DataFrame data using ECharts
  * library.
@@ -10,7 +15,12 @@ public class ECharts {
      * Starts a builder for a new chart.
      */
     public static EChart chart() {
-        return new EChart();
+        return chart(ElementIdGenerator.random());
+    }
+
+    // intentionally non-public for now. Any reason to expose ID generator API?
+    static EChart chart(ElementIdGenerator idGenerator) {
+        return new EChart(idGenerator);
     }
 
     /**
@@ -18,5 +28,30 @@ public class ECharts {
      */
     public static EChart chart(String title) {
         return chart().title(title);
+    }
+
+    /**
+     * @since 2.0.0
+     */
+    public static EChartHtmlSaver saver() {
+        return new EChartHtmlSaver();
+    }
+
+    public void save(Path path, EChartHtml... charts) {
+        saver().save(path, charts);
+    }
+
+    /**
+     * @since 2.0.0
+     */
+    public void save(File file, EChartHtml... charts) {
+        saver().save(file, charts);
+    }
+
+    /**
+     * @since 2.0.0
+     */
+    public void save(Appendable out, EChartHtml... charts) {
+        saver().save(out, charts);
     }
 }