+++ RELEASE-NOTES.md
@@ -7,6 +7,7 @@
 * #377 Upgrade SLF4J to 2.0.16
 * #378 Upgrade commons-compress to 1.27.1
 * #379 Upgrade Apache POI to 5.3.0
+* #384 EChartHtmlSaver to save charts to HTML
 
 ## Release 1.1.0